backend/requirements.txt
litestar[standard]==2.18.0
piccolo[postgres]>=1.0.0
pydantic>=2.0.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
# Auth
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
# Rate limiting
slowapi>=0.1.9
